feat: add QR code generator for local custom texts to guess
