+++ lib/features/categories/local/texts/texts.list.widget.dart
@@ -1,18 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:guess_the_text/features/categories/local/texts/edit.text.widget.dart';
-import 'package:guess_the_text/features/game/challenge/on.the.fly.challenge.model.dart';
+import 'package:guess_the_text/features/challenge/on.the.fly.challenge.model.dart';
 import 'package:guess_the_text/service.locator.dart';
 import 'package:guess_the_text/services/text.service/api.category.model.dart';
 import 'package:guess_the_text/services/text.service/api.text.model.dart';
 import 'package:guess_the_text/services/text.service/sql.db.service.dart';
 import 'package:guess_the_text/theme/app.theme.dart';
+import 'package:guess_the_text/theme/theme.utils.dart';
 import 'package:guess_the_text/theme/widgets/app.bar.title.widget.dart';
+import 'package:guess_the_text/theme/widgets/full.screen.bg.image.widget.dart';
 import 'package:guess_the_text/theme/widgets/snackbar/snackbar.model.dart';
 import 'package:guess_the_text/theme/widgets/snackbar/snackbar.utils.dart';
-
-import '/theme/theme.utils.dart';
-import '/theme/widgets/full.screen.bg.image.widget.dart';
 
 class TextsListWidget extends StatefulWidget {
   static const String backgroundImage = 'assets/images/backgrounds/background-pexels-pixabay-461940.jpg';
