+++ lib/route.generator.dart
@@ -1,13 +1,14 @@
 import 'package:flutter/material.dart';
+import 'package:guess_the_text/features/about/about.screen.dart';
+import 'package:guess_the_text/features/categories/categories.screen.dart';
+import 'package:guess_the_text/features/challenge/on.the.fly.chalenge.qr.screen.dart';
+import 'package:guess_the_text/features/challenge/on.the.fly.challenge.model.dart';
+import 'package:guess_the_text/features/game/game.screen.dart';
+import 'package:guess_the_text/features/settings/settings.screen.dart';
+import 'package:guess_the_text/service.locator.dart';
+import 'package:guess_the_text/services/logger/logger.service.dart';
 
-import '/features/about/about.screen.dart';
-import '/features/categories/categories.screen.dart';
-import '/features/game/challenge/on.the.fly.chalenge.qr.screen.dart';
-import '/features/game/challenge/on.the.fly.challenge.model.dart';
-import '/features/game/game.screen.dart';
-import '/features/settings/settings.screen.dart';
-import '/service.locator.dart';
-import '/services/logger/logger.service.dart';
+import 'features/challenge/on.the.fly.chalenge.qr.screen.dart';
 
 Route? onGenerateRoute(RouteSettings settings) {
   final LoggerService logger = serviceLocator.get();
